+++ app/javascript/controllers/map/layers_controller.js
@@ -7,13 +7,6 @@ import { draw } from 'maplibre/edit'
 import { status } from 'helpers/status'
 
 export default class extends Controller {
-  connect () {
-    // initializeMaplibreProperties is not called yet when rendering _share.haml
-    let props = mapProperties || window.gon.map_properties
-    document.querySelector('#map-view-permissions').value = props['view_permission']
-    document.querySelector('#map-edit-permissions').value = props['edit_permission']
-  }
-
   upload () {
     const fileInput = document.getElementById('fileInput')
     const file = fileInput.files[0]
@@ -96,16 +89,6 @@ export default class extends Controller {
       curve: 0.3,
       essential: true
     })
-  }
-
-  updateEditPermissions () {
-    mapProperties['edit_permission'] = document.querySelector('#map-edit-permissions').value
-    mapChannel.send_message('update_map', { edit_permission: mapProperties['edit_permission'] })
-  }
-
-  updateViewPermissions () {
-    mapProperties['view_permission'] = document.querySelector('#map-view-permissions').value
-    mapChannel.send_message('update_map', { view_permission: mapProperties['view_permission'] })
   }
 
   close () {