+++ spec/features/map_share_spec.rb
@@ -11,7 +11,7 @@ describe 'Map' do
     end
 
     it 'has share public link' do
-      expect(page).to have_link('Viewer link', href: '/m/' + subject.public_id)
+      expect(page).to have_link('View link', href: '/m/' + subject.public_id)
     end
 
     it 'has share private link' do
@@ -19,15 +19,15 @@ describe 'Map' do
     end
 
     it 'has share geojson link' do
-      expect(page).to have_link('Download geojson', href: '/m/' + subject.public_id + '.geojson')
+      expect(page).to have_link('GeoJSON', href: '/m/' + subject.public_id + '.geojson')
     end
 
     it 'has share gpx link' do
-      expect(page).to have_link('Download gpx', href: '/m/' + subject.public_id + '.gpx')
+      expect(page).to have_link('GPX', href: '/m/' + subject.public_id + '.gpx')
     end
 
     it 'has share map export link' do
-      expect(page).to have_link('Download map export', href: '/m/' + subject.public_id + '.json')
+      expect(page).to have_link('Map export', href: '/m/' + subject.public_id + '.json')
     end
   end
 
@@ -44,6 +44,25 @@ describe 'Map' do
 
     it 'can download gpx export' do
       visit '/m/' + subject.public_id + '.gpx'
+    end
+  end
+
+  context 'permissions' do
+    before do
+      visit map_path(map)
+      expect(page).to have_css("#maplibre-map[map-loaded='true']")
+      find('.maplibregl-ctrl-share').click
+      expect(page).to have_text('Share Map')
+    end
+
+    it 'can update view permission' do
+      select('Only you', from: 'map-view-permissions')
+      wait_for { map.reload.view_permission }.to eq('private')
+    end
+
+    it 'can update edit permission' do
+      select('Only you', from: 'map-edit-permissions')
+      wait_for { map.reload.edit_permission }.to eq('private')
     end
   end
 end